Add enough tests to get to 95% coverage
And, oh look, it's another bug in substitute fixed!

Doesn't seem like it's possible to add tests for DEBUG=1 coverage;
attempted it, and the tests definitely run but coverage doesn't notice
for some reason (or the coverage for those were lost). Not terribly
important though as long as all the non-debug functionality is covered.

Finally, move ut namespace down so main's at the top. Makes it a bit
easier to spot the "how to run the test suite" comments and fix the
name of the root element in the error messages.