Adds a test case for the new format under WSL 1
WSL 1 is still popular for some niche cases.
With this test case we assert that the image and its provisioning
command behaves well under such platform.
Ofc not all features are supported there, we just aim to assert that we
don't block forever waiting on systemd services and other assumptions
that wouldn't hold true in WSL 1.