A test case for WSL 2 and cloud-init
Also assuming the tar-based format, now distributed in
cdimages.ubuntu.com.

This particular test case verifies the interaction with cloud-init. If
it creates a non-system user, the provisioning workflow should skip
prompting for user creation. Other parts of the workflow should just
work.

This also verifies that the Ubuntu colors and font are applied to the
Windows terminal.